+++ zikzak_inappwebview_macos/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -130,6 +130,21 @@ class MacOSInAppWebViewController extends PlatformInAppWebViewController {
               NavigationActionPolicy.CANCEL.toNativeValue();
         }
         return NavigationActionPolicy.ALLOW.toNativeValue();
+      case 'onCreateWindow':
+        if (params.webviewParams?.onCreateWindow != null) {
+          Map<String, dynamic> arguments = call.arguments
+              .cast<String, dynamic>();
+          CreateWindowAction? createWindowAction = CreateWindowAction.fromMap(
+            arguments,
+          );
+          if (createWindowAction != null) {
+            return await params.webviewParams!.onCreateWindow!(
+              controller,
+              createWindowAction,
+            );
+          }
+        }
+        break;
       case 'onConsoleMessage':
         if (params.webviewParams?.onConsoleMessage != null) {
           var map =